Add support for the SDL mixer.
This is the only way to get audio support in Rust at the moment; the
low-level SDL_audio stuff wants to run a callback on a separate thread and that
will segfault at the moment. brson's stuff should eventually fix this, but for
now the SDL mixer must be used.