A variant of sdl-demo/main.rs that is designed to be used with the platform-specific SDL_main C/C++/Objective-C sources that take care of requirements that have to be taken care of on the main-thread before handing off control to the user's event loop.
Fix #118.

I tried to include documentation of how to actually build this thing,
but its probably not the ideal method.

Anyway, as far as I can tell, this is a closer match to the intent of
the SDL-distributed demos on win32 and Mac OS that redefine `main` as
`SDL_main` and define their own `main`; see the preprocessor
directives in `SDL_main.h`.